+++ tests/testapp/settings.py
@@ -52,7 +52,3 @@ MIDDLEWARE = [
     "django.contrib.messages.middleware.MessageMiddleware",
     "django.middleware.clickjacking.XFrameOptionsMiddleware",
 ]
-
-CONSCIOUS_EMBED_PROVIDERS = {
-    "example.com": "https://example.com/privacy/",
-}